requirements.txt
fastapi
pymongo
uvicorn
python-telegram-bot
python-dotenv
requests 
subprocess.run
pydantic
asyncio
openai
passlib
python-jose
google-generativeai
Pillow
ipython